Added middle layer class for functional tests to work with Mongo module.
